requirements.txt
openai-agents==0.0.15
Flask==3.1.1
Flask_Cors==5.0.0
pydantic==2.11.7
python-dotenv==1.1.0
Requests==2.32.4
